// src/app/shared/models/user.model.ts
 



export interface User {

    id?: number
  
    name?: string
    lastname?: string
    username?: string
    password?: string
    email?: string
    dob?: Date
    photoprofile?: string
    headline?: string
    currentPosition?: string
    education?: string
    location?: string
    contact_info?: string
    experience?: string
    points?: number
    culturalPoints?: number
    sportPoints?: number
    artPoints?: number
    professionelPoints?: number
    resetpasswordtoken?: string
    
}